Move TensorAccessorArgs to metal
